Check liveness of container and return log

+++ docker/entrypoint.sh
@@ -9,6 +9,10 @@ MEM_GB=$(printf %.1s $(free -m | tail -n+2 | head -n-1 | awk -F' ' '{ print $2 }
 
 if [ ! -z $IRI_DB_URL ]; then
   wget $IRI_DB_URL -O /tmp/testnetdb.tgz
+  if [ $(sha256sum /tmp/testnetdb.tgz | cut -d' ' -f1) != $IRI_DB_CHECKSUM ]; then
+    echo "ERROR: checksum $IRI_DB_CHECKSUM doesn't match downloaded file!" >&2
+    exit 2
+  fi
   rm -rf /iri/data/testnet*
   tar xfv /tmp/testnetdb.tgz --strip-components=1 -C /iri/data testnet_files/testnetdb
 fi
